+++ pkg/core/object/object.go
@@ -1,9 +1,14 @@
 package object
 
 import (
+	"errors"
+
 	"github.com/nspcc-dev/neofs-sdk-go/object"
 	oid "github.com/nspcc-dev/neofs-sdk-go/object/id"
 )
+
+// ErrInvalidSearchQuery is returned when some object search query is invalid.
+var ErrInvalidSearchQuery = errors.New("invalid search query")
 
 // AddressOf returns the address of the object.
 func AddressOf(obj *object.Object) oid.Address {