+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog for NeoFS Node
 ## [Unreleased]
 
 ### Added
+- Support of numeric object search queries (#2733)
 
 ### Fixed
 - Access to `PUT` objects no longer grants `DELETE` rights (#2261)